Release SP30 (#205)
* add await to signin

* add await to signin

* upd TokenService to use issuer token_endpoint

* add debug for token svc

* finish adding token endpoint to Token service

* add server-side rendering API ROOT

* refactor the keycloak services

* refactor the keycloak services

* refactor the keycloak services

* fix crashing bug with delete access

* refactor the issuer and uma2 services, updates to whitelist

* upd cicd deploy with SSR_API_ROOT

* Add queries to whitelist

* update whitelist

* update whitelist

* upd whitelist and permissions

* upd UI for environment plugins

* upd to env plugins for jwt-keycloak template

* upd whitelist

* better handling of my access requests

* refactor the access mgr and credential admin

* adj credential issuer retrieval

* upd whitelist

* cleanup poc manager pages

* upd auth profile failure messages

* fix error on environment detail

* adj authz matrix

* upd whitelist and upd business profile handling

* upd scope retrieval and consumer list

* fix service account list query

* upd whitelists

+++ src/services/workflow/delete-access.ts
@@ -10,6 +10,7 @@ import {
   KeycloakClientRegistrationService,
   KeycloakTokenService,
   getOpenidFromIssuer,
+  getUma2FromIssuer,
 } from '../keycloak';
 import { KongConsumerService } from '../kong';
 import { IssuerEnvironmentConfig, getIssuerEnvironmentConfig } from './types';
@@ -109,12 +110,13 @@ export const DeleteAccess = async (context: any, operation: any, keys: any) => {
             svc.productEnvironment.name
           );
           const openid = await getOpenidFromIssuer(issuerEnvConfig.issuerUrl);
+          const uma2 = await getUma2FromIssuer(issuerEnvConfig.issuerUrl);
           const token =
             issuerEnvConfig.clientRegistration == 'anonymous'
               ? null
               : issuerEnvConfig.clientRegistration == 'managed'
               ? await new KeycloakTokenService(
-                  openid.issuer
+                  openid.token_endpoint
                 ).getKeycloakSession(
                   issuerEnvConfig.clientId,
                   issuerEnvConfig.clientSecret
@@ -124,19 +126,28 @@ export const DeleteAccess = async (context: any, operation: any, keys: any) => {
           // Delete the Policies that are associated with the client!!
           // Use the kcprotect service to find the UMA Policies that have this client ID
           // and then delete each one
-          logger.debug('Deleting policies for %s', svc.consumer.customId);
-          const policyApi = new UMAPolicyService(openid.issuer, token);
-          const relatedPolicies = await policyApi.listPolicies({
-            name: svc.consumer.customId,
-          });
-          for (const policy of relatedPolicies.filter(
-            (p) => p.name == svc.consumer.customId
-          )) {
-            await policyApi.deleteUmaPolicy(policy.id);
+          // Only do the below if this authorization profile has UMA resources
+          if (issuer.resourceType != null && issuer.resourceType != '') {
+            logger.debug('Deleting policies for %s', svc.consumer.customId);
+            const policyApi = new UMAPolicyService(uma2.policy_endpoint, token);
+            const relatedPolicies = await policyApi.listPolicies({
+              name: svc.consumer.customId,
+            });
+            for (const policy of relatedPolicies.filter(
+              (p) => p.name == svc.consumer.customId
+            )) {
+              await policyApi.deleteUmaPolicy(policy.id);
+            }
+          } else {
+            logger.debug(
+              'Non-UMA Profile so skipping policy cleanup for %s',
+              svc.consumer.customId
+            );
           }
 
           await new KeycloakClientRegistrationService(
-            openid.issuer,
+            issuerEnvConfig.issuerUrl,
+            openid.registration_endpoint,
             token
           ).deleteClientRegistration(svc.consumer.customId);
         }
